+++ FakeNFT/Services/Requests/ProfileRequest.swift
@@ -8,11 +8,40 @@
 import Foundation
 
 struct ProfileRequest: NetworkRequest {
-    let endpoint: URL?
-    let dto: Dto?
+    var profile: Profile?
 
-    init() {
-        self.endpoint = URL(string: "\(RequestConstants.baseURL)/api/v1/profile/1")
-        self.dto = nil
+    var endpoint: URL? {
+        URL(string: "\(RequestConstants.baseURL)/api/v1/profile/1")
+    }
+
+    var httpMethod: HttpMethod = .put
+
+    var dto: Dto? {
+        guard
+            let profile,
+            let profileDTO = profile.toResponse()
+        else {
+            return nil
+        }
+
+        return UpdateProfileDto(profile: profileDTO)
+    }
+}
+
+struct UpdateProfileDto: Dto {
+    let profile: ProfileDTO
+
+    enum CodingKeys: String, CodingKey {
+        case likes
+        case avatar
+        case name
+    }
+
+    func asDictionary() -> [String: String] {
+        [
+            CodingKeys.likes.rawValue: profile.likes.joined(separator: ","),
+            CodingKeys.avatar.rawValue: profile.avatar,
+            CodingKeys.name.rawValue: profile.name
+        ]
     }
 }
